requirements.txt
fastapi==0.115.12
uvicorn[standard]
python-dotenv==1.1.0
pre-commit==4.2.0
